Move CLI-specific code to new `augur_cli` package
This cleans up the main `augur` namespace and better separates the
concerns of a library and a CLI tool.
